Favs view message empty and stast ui optimization

+++ Trivia App/Features/FavoritesView/FavoritesView.swift
@@ -24,37 +24,54 @@ struct FavoritesView: View {
             
             Spacer()
             
-            List {
-                ForEach(viewModel.questions) { question in
-                    DisclosureGroup {
-                        FavoritesDetailCardView(question: question)
-                            .listRowInsets(.init(top: 0, leading: -20, bottom: 0, trailing: 15))
-                    } label: {
-                        FavoritesCardView(question: question)
+            if(!viewModel.questions.isEmpty) {
+                List {
+                    ForEach(viewModel.questions) { question in
+                        DisclosureGroup {
+                            FavoritesDetailCardView(question: question)
+                                .listRowInsets(.init(top: 0, leading: -20, bottom: 0, trailing: 15))
+                        } label: {
+                            FavoritesCardView(question: question)
+                        }
+                        .tint(LinearGradient(
+                            colors: Gradients.redGradient,
+                            startPoint: .leading,
+                            endPoint: .trailing)
+                        )
+                        .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 5, trailing: 0))
+                        .listRowBackground(Color.clear)
+                        
                     }
-                    .tint(LinearGradient(
-                        colors: Gradients.redGradient,
-                        startPoint: .leading,
-                        endPoint: .trailing)
-                    )
-                    .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 5, trailing: 0))
-                    .listRowBackground(Color.clear)
+                    .onDelete(perform: { indexSet in
+                        Task {
+                            for index in indexSet {
+                                await viewModel.removeFavoriteQuestions(index: index)
+                            }
+                        }
+                    })
                     
                 }
-                .onDelete(perform: { indexSet in
-                    Task {
-                        for index in indexSet {
-                            await viewModel.removeFavoriteQuestions(index: index)
-                        }
-                    }
-                })
+                .listRowSpacing(10)
+                .scrollContentBackground(.hidden)
                 
+                Spacer()
+            } else {
+                VStack(alignment: .center) {
+                    Spacer()
+                    
+                    Text("favoritesView.favoritesQuestionsEmpty".localized())
+                        .font(.title2)
+                        .fontWeight(.light)
+                        .padding(.vertical)
+                    
+                    Image("question")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(maxWidth: 130)
+                    
+                    Spacer()
+                }
             }
-            .listRowSpacing(10)
-            .scrollContentBackground(.hidden)
-            
-            Spacer()
-            
         }.setBackgroundApp()
         .task {
             await viewModel.getFavoriteQuestions()
